Add files via upload
complete

+++ controller/userController.js
@@ -1,18 +1,21 @@
-import cloudinary from "../config/cloudinary.js";
 import HttpError from "../middleware/HttpError.js";
-import uploads from "../middleware/upload.js";
-import User from "../model/userModel.js";
+import User from "../model/User.js";
+import cloudinary from "../config/cloudinary.js";
+import sendEmail from "../utils/sendEmail.js";
+import { getWelcomeEmailTemplate, getResetPasswordTemplate } from "../services/emailTemplate.js"
+
+import crypto from "crypto";
 
 const add = async (req, res, next) => {
   try {
-    const { name, email, password, phone, role } = req.body;
+    const { name, email, password, role, phone } = req.body;
 
     const newUser = {
       name,
       email,
       password,
-      phone,
       role,
+      phone,
       profilePic: req.file ? req.file.path : "undefined",
       cloudinaryId: req.file ? req.file.filename : "undefined",
     };
@@ -23,9 +26,15 @@ const add = async (req, res, next) => {
 
     await user.save();
 
+    sendEmail({
+      to: newUser.email,
+      subject: "welcome to QuickNest",
+      html: getWelcomeEmailTemplate(newUser.name)
+    })
+
     res.status(201).json({ success: true, user });
   } catch (error) {
-    next(error);
+    next(new HttpError(error.message, 500));
   }
 };
 
@@ -38,14 +47,12 @@ const login = async (req, res, next) => {
     const token = await user.generateAuthToken();
 
     if (!user) {
-      return next(new HttpError("Unable to login ", 400));
+      return next(new HttpError("unable to login"));
     }
 
-    res
-      .status(200)
-      .json({ success: true, message: "Login Successfully", user, token });
+    res.status(200).json({ success: true, user, token });
   } catch (error) {
-    next(new HttpError(error.message, 404));
+    next(new HttpError(error.message, 500));
   }
 };
 
@@ -54,29 +61,26 @@ const authLogin = async (req, res, next) => {
     const user = req.user;
 
     if (!user) {
-      return next(new HttpError("Unable to login"));
+      return next(new HttpError("user not found", 404));
     }
 
-    res.status(201).json({ success: true, user });
+    res.status(200).json({ success: true, user });
   } catch (error) {
-    next(new HttpError(error.message, 404));
+    next(new HttpError(error.message, 500));
   }
 };
 
 const logOut = async (req, res, next) => {
   try {
-    const token = req.token; 
-
     req.user.tokens = req.user.tokens.filter((t) => {
-      return t.token !== token;
+      return t.token != req.token;
     });
 
     await req.user.save();
 
-    res.status(200).json({
-      success: true,
-      message: "User LogOut Successfully",
-    });
+    res
+      .status(200)
+      .json({ success: true, message: "user logOut successfully" });
   } catch (error) {
     next(new HttpError(error.message, 500));
   }
@@ -88,33 +92,55 @@ const logOutAll = async (req, res, next) => {
 
     await req.user.save();
 
-    res
-      .status(200)
-      .json({ success: true, message: "User Logout from all device" });
+    res.status(200).json({
+      success: true,
+      message: "user logOut from all device successfully",
+    });
   } catch (error) {
-    next(new HttpError(error.message, 404));
+    next(new HttpError(error.message, 500));
   }
 };
 
-const getAll = async (req, res, next) => {
+const allUser = async (req, res, next) => {
   try {
-    const users = await User.find({});
 
-    if (users.length === 0) {
-      res
-        .status(200)
-        .json({ success: true, message: "No any users are there" });
+    const { role, limit, skip, sortBy } = req.query
+
+    let query = {};
+
+    let sortByValue = {};
+
+    if (role) {
+      query.role = role
     }
 
-    res.status(200).json({ success: true, message: "all users", users });
+    if (sortBy) {
+
+      const [field, order] = sortBy.split(":");
+
+      sortByValue[field] = order === "desc" ? -1 : 1;
+
+    }
+
+    const users = await User.find(query).limit(parseInt(limit) || 5).skip(parseInt(skip) || 0).sort(sortByValue);
+
+    if (users.length === 0) {
+      res.status(200).json({ success: true, message: "no user data found" });
+    }
+
+    res
+      .status(200)
+      .json({ success: true, message: "all user data fetched", length: users.length, users });
   } catch (error) {
-    next(new HttpError(error.message, 404));
+    next(new HttpError(error.message, 500));
   }
 };
 
 const update = async (req, res, next) => {
   try {
-    const user = req.user;
+    let targetedUser = req.params.id || req.user._id;
+
+    const user = await User.findById(targetedUser);
 
     if (!user) {
       return next(new HttpError("user not found", 404));
@@ -122,18 +148,32 @@ const update = async (req, res, next) => {
 
     const updates = Object.keys(req.body);
 
-    const allowedField = ["name", "password", "phone"];
+    let allowedFields = ["name", "password", "phone", "profilePic"];
 
-    const isValid = updates.every((field) => allowedField.includes(field));
+    if (req.user.role === "admin" || req.user.role === "super_admin") {
+      allowedFields = [...allowedFields, "role", "isVerified"];
+    }
+
+    const isValid = updates.every((field) => allowedFields.includes(field));
 
     if (!isValid) {
-      return next(new HttpError("only allowed fields can be updated", 400));
+      return next(new HttpError("only allowed field can be updated", 400));
+    }
+
+    if (
+      !req.user.role === "admin" &&
+      !req.user.role === "super_admin" &&
+      !req.user._id.toString() !== user._id.toString()
+    ) {
+      return next(new HttpError("unauthorized access", 401));
     }
 
     updates.forEach((update) => (user[update] = req.body[update]));
 
     if (req.file) {
-      await cloudinary.uploader.destroy(user.cloudinaryId);
+      if (user.cloudinaryId) {
+        await cloudinary.uploader.destroy(user.cloudinaryId);
+      }
 
       user.profilePic = req.file.path;
 
@@ -144,27 +184,138 @@ const update = async (req, res, next) => {
 
     res
       .status(200)
-      .json({ success: true, message: "user update successfully", user });
+      .json({ success: true, message: "user Data updated successfully", user });
   } catch (error) {
-    next(new HttpError(error.message, 404));
+    next(new HttpError(error.message));
   }
 };
 
 const deleteUser = async (req, res, next) => {
   try {
-    const user = req.user;
+    const targetedUser = req.params.id || req.user._id;
+
+    const user = await User.findById(targetedUser);
+
+    if (!user) {
+      return next(new HttpError("user not found"), 401);
+    }
+
+    if (
+      !req.user.role === "admin" &&
+      !req.user.role === "super_admin" &&
+      !req.user._id.toString() !== user._id.toString()
+    ) {
+      return next(new HttpError("unauthorized access", 401));
+    }
 
     await User.deleteOne(user);
 
-    await cloudinary.uploader.destroy(user.cloudinaryId);
+    if (user.cloudinaryId) {
+      await cloudinary.uploader.destroy(user.cloudinaryId);
+    }
 
     res
       .status(200)
-      .json({ success: true, message: "user delete successfully" });
+      .json({ success: true, message: "user deleted successfully" });
   } catch (error) {
     next(new HttpError(error.message, 500));
   }
 };
+
+
+const forgotPassword = async (req, res, next) => {
+
+  try {
+
+    const { email } = req.body;
+
+    const user = await User.findOne({ email })
+
+
+    if (!user) {
+
+      return next(new HttpError("user not found", 404))
+
+    }
+
+    const resetToken = crypto.randomBytes(32).toString("hex");
+
+    const hashedToken = crypto.createHash("sha256").update(resetToken).digest("hex");
+
+
+    user.resetPasswordToken = hashedToken;
+    user.resetPasswordExPiry =  Date.now() + 15 * 60 * 1000;
+
+
+    await user.save();
+
+
+    const resetLink = `localhost:5000/user/reset-password/${resetToken}`
+
+
+
+    await sendEmail({
+      to: user.email,
+      subject: "Password reset Request",
+      html: getResetPasswordTemplate(user.name, resetLink),
+    })
+
+
+    res.status(200).json({ success: true, message: "password reset link send to email successfully", resetLink });
+
+
+
+  } catch (error) {
+
+    next(new HttpError(error.message))
+
+  }
+
+}
+
+
+const resetPassword = async (req, res, next) => {
+
+  try {
+
+    const { token } = req.params;
+
+
+    const { newPassword, confirmPassword } = req.body;
+
+    if (newPassword !== confirmPassword) {
+      return next(new HttpError("password is not matched", 400));
+
+    }
+
+
+    const hashedToken = crypto.createHash("sha256").update(token).digest("hex")
+
+
+    const user = await User.findOne({
+      resetPasswordToken: hashedToken,
+      resetPasswordExPiry: { $gt: Date.now() }
+
+    })
+
+
+    if (!user) {
+      return next(new HttpError("password or token is expired please try again", 400))
+    }
+
+    user.password = confirmPassword;
+    user.resetPasswordToken = null,
+      user.resetPasswordExPiry = null
+
+    await user.save();
+
+
+    res.status(200).json({ success: true, message: "password update successfully" })
+
+  } catch (error) {
+    next(new HttpError(error.message))
+  }
+}
 
 export default {
   add,
@@ -172,7 +323,9 @@ export default {
   authLogin,
   logOut,
   logOutAll,
-  getAll,
+  allUser,
   update,
   deleteUser,
+  forgotPassword,
+  resetPassword
 };
